Reworking online game starter
